Revert rename search route
